pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "homewakeword-bcresnet"
version = "0.0.0"
description = "HomeWakeWord BC-ResNet Home Assistant add-on skeleton"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6",
]

[tool.setuptools]
packages = ["homewakeword"]

[tool.pytest.ini_options]
testpaths = ["tests"]
